requirements.txt
streamlit
openai
sqlalchemy
pymysql
python-dotenv
pytz